Upload completed tip.py

+++ 0-FunctionsVariables/calculator.py
@@ -1,6 +1,6 @@
-x = int(input("Enter x: "))  # Ask the user for x
-y = int(input("Enter y: "))  # Ask the user for y
-z = int(input("Enter z: "))  # Ask the user for z
+x = float(input("Enter x: "))  # Ask the user for x
+y = float(input("Enter y: "))  # Ask the user for y
+z = float(input("Enter z: "))  # Ask the user for z
 
 print(x + y + z) # Print the sum of x, y, and z
 print(x - y - z) # Print the difference of x, y, and z
